Move `TaskContext` to `datafusion_execution`
